PHP Fatal error:  Uncaught Error: Unsupported operand types
If an item has no rating yet the value of $theme['rating']['count'] /$theme['rating'] is null instead of 0.
This results in a fatal error.

So you have to check if $theme['rating'] is an array and the key count exists.
I assume that if $theme['rating'] is an array then the key "count" should be set.
Then on line 191 I check if the value is greater than zero to skip an unnecessary math operation because any math operation with 0 leads to 0. If the value is null it will be cast to 0 and therefore this will not throw a fatal error any longer.
If the key count doesn't exists (which should if $theme['rating'] is an array) or $theme['rating'] isn't an array the the code will be executed after lie 196. 
Here I also check the value to skip the unnecessary math operation.